aws/rds/requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
